# api-contract-v1.yaml
openapi: 3.1.0
info:
  title: Disaster Early-Warning API
  version: 0.5.0
  description: >
    Contract v1. Region-aware by design; US is the only enabled region in v1.

    CHANGELOG
    0.5.0 — Reports are write-first (fusion runs after the 202):
      ~ ReportAccepted: event_id and tier are now optional (set once fusion has run);
        + report_id, + status (pending | fused | failed)
      + GET /reports/{client_event_id} (the reporter's own report status)
      + 409 on POST /reports when the client_event_id belongs to another reporter; 429 when rate limited
    0.4.0 — Additive only, no breaking changes (per-user alert preferences, marine alerts):
      + AlertPreferences schema; User.alert_preferences
      + PUT /me/alert-preferences
      + GET /alerts: hazard_types, min_severity, include_marine (default to the caller's
        saved preferences when signed in)
      + Alert.product, Alert.marine
    0.3.0 — Additive only, no breaking changes (fields the mobile alert card needs):
      + Alert.hazard_type          (drives the card's hazard icon)
      + Alert.location_name        (place shown in the card's metadata line)
      + Alert.corroboration_count  ("Confirmed by N nearby" on corroborated reports)
    0.2.0 — Additive only, no breaking changes:
      + GET /events/{id}/brief        (Ultra situational brief, async)
      + GET /regions/{region}/cascade (Ultra cascade analysis)
      + GET /guidance/manifest        (offline card sync)
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /events/{event_id}:
    get:
      tags: [reports]
      summary: Fused event detail
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Event' }
        '404': { $ref: '#/components/responses/NotFound' }
  /events/{event_id}/reports:
    get:
      tags: [reports]
      summary: Community reports fused into an event (public, coarse)
      description: >
        For the alert detail screen. Never includes notes, reporter identity or exact
        positions — only distance from the event, rounded to 100 m.
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventReports' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /events/{event_id}/brief:
    get:
      tags: [analysis]
      summary: Situational brief for an event
      description: >
        Generated asynchronously on Ultra after the fast alert has already been sent.
        Returns 202 with status "pending" while generation is in flight — clients
        should render the alert immediately and let the brief arrive behind it.
        Never block alert display on this endpoint.
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Brief ready
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SituationalBrief' }
        '202':
          description: Generation in flight
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [pending] }
                  retry_after_seconds: { type: integer }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /alerts:
    get:
      tags: [alerts]
      summary: Active alerts
      description: >
        Returns Tier 2 (public) alerts by default. Lower tiers are visible only to
        reviewers, or to subscribers who opt into early/unverified signals.
      parameters:
        - name: region
          in: query
          schema: { $ref: '#/components/schemas/RegionCode' }
        - name: lat
          in: query
          schema: { type: number, format: double }
        - name: lon
          in: query
          schema: { type: number, format: double }
        - name: radius_km
          in: query
          schema: { type: number, default: 50, maximum: 500 }
        - name: min_tier
          in: query
          schema: { type: integer, enum: [0, 1, 2], default: 2 }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: hazard_types
          in: query
          description: Only these hazard types. Defaults to the caller's saved preferences, else all.
          style: form
          explode: false
          schema:
            type: array
            items: { $ref: '#/components/schemas/HazardType' }
        - name: min_severity
          in: query
          description: Defaults to the caller's saved preferences, else low.
          schema: { $ref: '#/components/schemas/Severity' }
        - name: include_marine
          in: query
          description: Include marine (offshore/boating) products. Defaults to the caller's saved preferences, else false.
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items: { $ref: '#/components/schemas/Alert' }
  /alerts/{alert_id}:
    get:
      tags: [alerts]
      summary: Alert detail
      parameters:
        - name: alert_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Alert' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------- subscriptions ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    RegionCode:
      type: string
      description: >
        Enabled regions. v1 ships US only; CN is reserved and rejected at runtime
        until its feed, coordinate-frame and push work lands.
      enum: [US, CN]
    CoordinateFrame:
      type: string
      description: >
        Declared per region so clients never assume WGS-84. v1 is always WGS84.
      enum: [WGS84, GCJ02]
    GeoPoint:
      type: object
      required: [lat, lon]
      properties:
        lat: { type: number, format: double, minimum: -90, maximum: 90 }
        lon: { type: number, format: double, minimum: -180, maximum: 180 }
        accuracy_m: { type: number }
        frame:
          allOf: [{ $ref: '#/components/schemas/CoordinateFrame' }]
          default: WGS84
    HazardType:
      type: string
      enum: [flood, fire, earthquake, storm, landslide, other]
    ObservedEffect:
      type: string
      description: Controlled vocabulary — lets fusion match reports of the same phenomenon.
      enum:
        - rising_water
        - structural_damage
        - smoke_or_fire_visible
        - ground_shaking
        - blocked_road
        - other
    Severity:
      type: string
      enum: [low, medium, high, critical]
    SourceType:
      type: string
      enum: [official, manual, mesh]
...
    Event:
      type: object
      required: [event_id, source, hazard_type, severity, location, tier, confidence]
      properties:
        event_id: { type: string }
        source: { $ref: '#/components/schemas/SourceType' }
        region: { $ref: '#/components/schemas/RegionCode' }
        hazard_type: { $ref: '#/components/schemas/HazardType' }
        observed_effect: { $ref: '#/components/schemas/ObservedEffect' }
        severity: { $ref: '#/components/schemas/Severity' }
        location: { $ref: '#/components/schemas/GeoPoint' }
        first_reported_at: { type: string, format: date-time }
        last_updated_at: { type: string, format: date-time }
        tier:
          type: integer
          enum: [0, 1, 2]
          description: 0 unverified, 1 corroborated, 2 confirmed (public alert)
        confidence: { type: number, minimum: 0, maximum: 1 }
        specificity_score: { type: number, minimum: 0, maximum: 1 }
        distinct_reporter_count:
          type: integer
          description: Distinct originators only — relay hops never increment this.
        confirm_count: { type: integer }
        deny_count: { type: integer }
        official_match_id:
          type: string
          nullable: true
          description: Matching NOAA/USGS/EONET/GDACS event, if any.
        reviewer_decision:
          type: string
          nullable: true
          enum: [approve, reject, hold]
        brief_status:
          type: string
          description: Situational brief generation state. Clients never block on this.
          enum: [not_requested, pending, ready, failed]
          default: not_requested
        verification_label:
          type: string
          description: Shown verbatim in UI — unverified reports must never render as official.
          enum: [unverified_report, corroborated_report, official_confirmed]
    Alert:
      type: object
      required: [alert_id, event_id, headline, severity, issued_at, verification_label]
      properties:
        alert_id: { type: string }
        event_id: { type: string }
        headline: { type: string, maxLength: 120 }
        body: { type: string, maxLength: 600 }
        severity: { $ref: '#/components/schemas/Severity' }
        location: { $ref: '#/components/schemas/GeoPoint' }
        affected_radius_km: { type: number }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        verification_label:
          type: string
          enum: [unverified_report, corroborated_report, official_confirmed]
        source_attribution:
          type: string
          description: e.g. "NOAA/NWS" or "Community reports"
        hazard_type:
          $ref: '#/components/schemas/HazardType'
          description: Same value as the fused Event. Clients fall back to a generic icon when absent.
        location_name:
          type: string
          maxLength: 80
          description: >
            Human-readable place for the alert's location, e.g. "Russian River Basin".
            Shown beside distance and time; never repeated in the headline.
        corroboration_count:
          type: integer
          minimum: 0
          description: >
            Distinct originating reporters behind a corroborated_report (relay hops and
            repeat reports from one device do not count — see escalation caution rules).
            Omitted for official_confirmed; 1 or omitted for unverified_report.
        product:
          type: string
          description: >
            The source's own product name, e.g. "Flood Warning", "Small Craft Advisory",
            "earthquake". Finer-grained than hazard_type.
        marine:
          type: boolean
          description: Offshore/boating product; absent means false. Hidden unless the user opts in.
    EventReports:
      type: object
      required: [event_id, distinct_reporter_count, reports]
      properties:
        event_id: { type: string }
        distinct_reporter_count:
          type: integer
          description: Distinct people behind the reports (repeat reports from one person count once).
        reports:
          type: array
          description: Newest first, at most 50.
          items: { $ref: '#/components/schemas/ReportSummary' }
    ReportSummary:
      type: object
      required: [hazard_type, severity, observed_at, distance_from_event_m]
      properties:
        hazard_type: { $ref: '#/components/schemas/HazardType' }
        observed_effect: { $ref: '#/components/schemas/ObservedEffect' }
        severity: { $ref: '#/components/schemas/Severity' }
        observed_at: { type: string, format: date-time }
        distance_from_event_m: { type: integer, description: Rounded to 100 m. }
        captured_offline: { type: boolean }
        via_mesh: { type: boolean, description: Arrived over the Bluetooth mesh. }
    SituationalBrief:
      type: object
      required: [event_id, generated_at, model, summary, uncertainty]
      properties:
        event_id: { type: string }
        generated_at: { type: string, format: date-time }
        model: { type: string, description: e.g. nemotron-3-ultra }
        summary:
          type: string
          description: What is likely happening, in plain language.
        likely_progression:
          type: string
          description: Plausible near-term development. Explicitly an estimate.
        exposed_areas:
          type: array
          description: Downstream or adjacent areas that may be affected.
          items:
            type: object
            properties:
              description: { type: string }
              location: { $ref: '#/components/schemas/GeoPoint' }
              rationale: { type: string }
        official_guidance:
          type: string
          description: Guidance drawn from matched official sources, attributed.
        uncertainty:
          type: string
          description: >
            Required. What the brief does not know and where it may be wrong.
            Clients MUST display this — never render a brief as certainty.
        sources:
          type: array
          items: { type: string }
...